fixed bug regarding BMI score showing 0

+++ app/src/main/java/com/example/fitnesscalc/Formulas.kt
@@ -3,15 +3,15 @@ package com.example.fitnesscalc
 class Formulas {
 
     fun calculateBMI(pounds: Int, feet: Int, inches: Int) : Float {
-        val finalHeight: Int = (feet * 12) + inches
-        val finalHeightSquared: Int = finalHeight * finalHeight
+        val finalHeight: Float = (feet.toFloat() * 12) + inches.toFloat()
+        val finalHeightSquared: Float = finalHeight * finalHeight
 
 
-        val finalBMI = (pounds/finalHeightSquared) * 703
+        val finalBMI = (pounds.toFloat()/finalHeightSquared) * 703
 
 
 
-        return finalBMI.toFloat()
+        return finalBMI
 
     }
 }
